requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
numpy==1.24.1
pandas==1.5.2
pybit==2.4.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
websocket-client==1.4.2
websockets==10.4
